feat: add reducers
