+++ to-do-list/src/state/state.js
@@ -1,7 +1,5 @@
-const UPDATE_INPUT = 'UPDATE-INPUT';
-const ADD_TO_DO = 'ADD-TO-DO';
-const CHANGE_TO_DO = 'CHANGE-TO-DO';
-const DELETE_TO_DO = 'DELETE-TO-DO';
+import addReducer from './add-reducer';
+import doReducer from './do-reducer';
 
 let store = {
     _state: {
@@ -37,54 +35,13 @@ let store = {
         this.reRender = observer;
     },
     dispatch(action) {
-        switch(action.type){
-            case ADD_TO_DO:
-                this._state.needToDo.push(
-                    {
-                        id: this._state.needToDo.length > 0 ? this._state.needToDo[this._state.needToDo.length - 1].id + 1 : 0,
-                        toDo: this._state.input,
-                        isDone: false
-                    }
-                )
-                this._state.input = '';
-                this.reRender(store);
-                break;
-            case DELETE_TO_DO:
-                this._state.needToDo = this._state.needToDo.filter(el => el.id !== action.id);
-                this.reRender(store);
-                break;
-            case UPDATE_INPUT:
-                this._state.input = action.text;
-                this.reRender(store);
-                break;
-            case CHANGE_TO_DO:
-                this._state.needToDo = this._state.needToDo.map(el => {
-                    if (el.id === action.id) {
-                        if (el.isDone) {
-                            el.isDone = false;
-                            return el;
-                        }
-                        else if (!el.isDone) {
-                            el.isDone = true;
-                            return el;
-                        }
-                    }
-                    else {
-                        return el;
-                    }
-        
-                })
-                console.log(this._state)
-                this.reRender(store);
-                break;
-        }
+        this._state = doReducer(this._state, action);
+        this._state = addReducer(this._state, action);
+        this.reRender(store);
     }
 
 }
-export const actionAddToDo = () => ({type: ADD_TO_DO});
-export const actionDeleteToDo = (id) => ({type: DELETE_TO_DO, id: id});
-export const actionUpdateInput = (text) => ({type: UPDATE_INPUT, text: text});
-export const actionToDOChange = (id) => ({type: CHANGE_TO_DO, id: id})
+
 /*let state;
 
 localStorage.getItem('state')?
